refactor(genui): export XML fragment utilities
