+++ packages/genui/server/test/html-fragment-to-main-thread-script-tool.test.ts
@@ -7,7 +7,6 @@ import { describe, expect, test } from '@rstest/core';
 import {
   createHtmlFragmentScriptRunScope,
   createHtmlFragmentToMainThreadScriptTool,
-  generateMainThreadScript,
   resolveHtmlFragmentScriptPlaceholders,
 } from '../agent/html-fragment-to-main-thread-script-tool.js';
 
@@ -30,91 +29,6 @@ async function executeFragmentTool(
 }
 
 describe('HTML fragment to main-thread script tool', () => {
-  test('generates ordered Element PAPI calls for nested Lynx elements', () => {
-    const javascript = generateMainThreadScript(`
-      <scroll-view class="feed" id="root" style="height: 100vh;" scroll-orientation="vertical">
-        <view class="card" data-kind="featured">
-          <text aria-label="Greeting">Hello &amp; welcome</text>
-          <image src="https://example.com/cover.png" />
-        </view>
-      </scroll-view>
-    `);
-
-    expect(javascript).toBe(`const node0 = __CreateScrollView(pageId);
-__SetClasses(node0, "feed");
-__SetID(node0, "root");
-__SetInlineStyles(node0, "height: 100vh;");
-__SetAttribute(node0, "scroll-orientation", "vertical");
-const node1 = __CreateView(pageId);
-__SetClasses(node1, "card");
-__AddDataset(node1, "kind", "featured");
-const node2 = __CreateText(pageId);
-__SetAttribute(node2, "aria-label", "Greeting");
-__AppendElement(node2, __CreateRawText("Hello & welcome"));
-__AppendElement(node1, node2);
-const node3 = __CreateImage(pageId);
-__SetAttribute(node3, "src", "https://example.com/cover.png");
-__AppendElement(node1, node3);
-__AppendElement(node0, node1);
-__AppendElement(page, node0);`);
-  });
-
-  test('supports multiple roots, text content, and generic element tags', () => {
-    expect(
-      generateMainThreadScript(
-        '<view>Before<text>inside</text>after</view><input value="42" />',
-      ),
-    ).toBe(`const node0 = __CreateView(pageId);
-const node1 = __CreateText(pageId);
-__AppendElement(node1, __CreateRawText("Before"));
-__AppendElement(node0, node1);
-const node2 = __CreateText(pageId);
-__AppendElement(node2, __CreateRawText("inside"));
-__AppendElement(node0, node2);
-const node3 = __CreateText(pageId);
-__AppendElement(node3, __CreateRawText("after"));
-__AppendElement(node0, node3);
-__AppendElement(page, node0);
-const node4 = __CreateElement("input", pageId);
-__SetAttribute(node4, "value", "42");
-__AppendElement(page, node4);`);
-  });
-
-  test('preserves meaningful text whitespace and ignores whitespace-only nodes', () => {
-    const javascript = generateMainThreadScript(
-      '<text>  spaced text  </text><view>   </view>',
-    );
-
-    expect(javascript).toContain('__CreateRawText("  spaced text  ")');
-    expect(javascript).not.toContain('__CreateRawText("   ")');
-  });
-
-  test('escapes closing script sequences in generated JavaScript', () => {
-    expect(generateMainThreadScript('<text>&lt;/script&gt;</text>')).toContain(
-      '__CreateRawText("<\\/script>")',
-    );
-  });
-
-  test('rejects empty and malformed XML fragments', () => {
-    expect(() => generateMainThreadScript('   ')).toThrow(
-      'XML fragment must not be empty',
-    );
-    expect(() => generateMainThreadScript('<view><text></view>')).toThrow(
-      'Invalid XML fragment',
-    );
-    expect(() => generateMainThreadScript('<script />')).toThrow(
-      'Element <script> is not allowed',
-    );
-    expect(() =>
-      generateMainThreadScript('<view id="duplicate"/><text id="duplicate"/>')
-    ).toThrow('Duplicate XML id: duplicate');
-
-    const overlyDeepFragment = `${'<view>'.repeat(65)}${'</view>'.repeat(65)}`;
-    expect(() => generateMainThreadScript(overlyDeepFragment)).toThrow(
-      'XML fragment must not exceed 64 levels of element nesting',
-    );
-  });
-
   test('returns an opaque placeholder and id-to-node bindings', async () => {
     const { output, scope } = await executeFragmentTool(
       '<view id="root"><text id="label">Hello</text></view>',
